+++ packages/sp-clientsvc/rollup.es2015.config.js
@@ -1,0 +1,15 @@
+const sourcemaps = require("rollup-plugin-sourcemaps");
+
+const moduleName = "sp-clientsvc";
+
+module.exports = {
+    input: `./build/packages/${moduleName}/index.js`,
+    plugins: [sourcemaps()],
+
+    output: {
+        file: `./dist/packages/${moduleName}/dist/${moduleName}.js`,
+        format: "es",
+        sourcemap: true,
+    }
+};
+